fix: Add missing BioFeedbackSystem enum cases and fix CameraPPGProcessor type errors
**Enum Cases Added:**
- BioFeedbackSystem.h: Added Auto + NetworkStream cases (2 switches)
- CameraPPGProcessor.h: Fixed jmax/jmin type mismatches with explicit casts

**Progress: 990 → 140 warnings (86% reduction) ✅**

**Remaining: 140 warnings**
- 96 -Wsign-conversion
- 21 -Wunused-parameter
- 19 -Wswitch-enum
- 3 -Wsign-compare
- 1 -Wunused-but-set-variable

**Build Status:** ✅ Compiles successfully

+++ Sources/BioData/CameraPPGProcessor.h
@@ -297,12 +297,12 @@ private:
             float sum = 0.0f;
             int count = 0;
 
-            int start = juce::jmax(0, i - windowSize / 2);
-            int end = juce::jmin(static_cast<int>(signal.size()), i + windowSize / 2);
+            int start = juce::jmax(0, static_cast<int>(i) - windowSize / 2);
+            int end = juce::jmin(static_cast<int>(signal.size()), static_cast<int>(i) + windowSize / 2);
 
             for (int j = start; j < end; ++j)
             {
-                sum += signal[j];
+                sum += signal[static_cast<size_t>(j)];
                 count++;
             }
